Refactor error middleware for testing and extension
